requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2
numpy>=1.24.0
scikit-learn>=1.3.0

# LangChain & AI
langgraph>=0.0.20
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
langchain-groq

# Azure OpenAI (for multilingual chatbot)
openai>=1.0.0

# Azure Speech Services (for multilingual chatbot)
azure-cognitiveservices-speech>=1.31.0

# Email
secure-smtplib>=0.1.1

# Forecasting (TimeGPT)
nixtla>=0.1.0

# AI Insights
groq>=0.4.0

# Optional: For better performance
aiofiles>=23.0.0